Move betty.plugins.media_type to betty.media_types
