+++ betty/media_types/svg.py
