+++ betty/plugins/extension/maps/jobs.py
@@ -9,8 +9,8 @@ from typing import TYPE_CHECKING, override
 
 from betty.file import write
 from betty.job import Job
+from betty.media_types.html import HTML
 from betty.plugins.entity.place import Place
-from betty.plugins.media_type.html import HTML
 
 if TYPE_CHECKING:
     from babel import Locale
