+++ betty/plugins/jinja_filter/image_resize_cover.py
@@ -20,11 +20,11 @@ from betty.hashid import hashid_file_meta
 from betty.image import image_file_path_format, resize_cover
 from betty.jinja import context_document
 from betty.jinja.filter import JinjaFilter, JinjaFilterDefinition
+from betty.media_types.svg import SVG
 from betty.os import _link_or_copy
 from betty.plugins.entity.file import File
 from betty.plugins.entity.file_reference import FileReference
 from betty.plugins.jinja_filter.file import File as FileFilter
-from betty.plugins.media_type.svg import SVG
 from betty.project import Project
 
 if TYPE_CHECKING:
