+++ betty/project/generate/jobs.py
@@ -19,9 +19,9 @@ from betty.jinja import make_copy_function
 from betty.job import Job
 from betty.locale.localizable.gettext import _
 from betty.locale.localize import DEFAULT_LOCALIZER
+from betty.media_types.html import HTML
+from betty.media_types.json import JSON
 from betty.openapi import Specification
-from betty.plugins.media_type.html import HTML
-from betty.plugins.media_type.json import JSON
 from betty.privacy.resolve import is_public
 from betty.project.schema import ProjectSchema
 from betty.string import kebab_case_to_lower_camel_case
