+++ betty/tests/project/test_url.py
@@ -8,8 +8,8 @@ from betty.entity import EntityDefinition
 from betty.entity.collection.pool import EntityPool
 from betty.locale import DEFAULT_LOCALE, DEFAULT_LOCALE_TAG, ResolvableLocale
 from betty.media_type import MediaType
-from betty.plugins.media_type.html import HTML
-from betty.plugins.media_type.json import JSON
+from betty.media_types.html import HTML
+from betty.media_types.json import JSON
 from betty.project import Project
 from betty.project.url import (
     _EntityUrlGenerator,
