+++ betty/tests/test_image.py
@@ -12,8 +12,8 @@ from betty.image import (
     resize_cover,
 )
 from betty.media_type import MediaType
-from betty.plugins.media_type.pdf import PDF
-from betty.plugins.media_type.svg import SVG
+from betty.media_types.pdf import PDF
+from betty.media_types.svg import SVG
 
 
 @pytest.mark.parametrize(
