+++ betty/tests/url/test___init__.py
@@ -6,7 +6,7 @@ import pytest
 from babel import Locale
 
 from betty.locale import DEFAULT_LOCALE, DEFAULT_LOCALE_TAG, ResolvableLocale
-from betty.plugins.media_type.html import HTML
+from betty.media_types.html import HTML
 from betty.url import PassthroughUrlGenerator, generate_from_path
 
 if TYPE_CHECKING:
